requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
contourpy==1.2.1
cycler==0.12.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.4
django-bootstrap4==24.3
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-import-export==3.3.8
django-widget-tweaks==1.5.0
et-xmlfile==1.1.0
ezdxf==1.2.0
fonttools==4.51.0
kiwisolver==1.4.5
MarkupPy==1.14
matplotlib==3.8.4
numpy==1.26.4
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tablib==3.5.0
typing_extensions==4.11.0
xlrd==2.0.1
xlwt==1.3.0
